clean up, add routing, organise components
