+++ src/MainLandingElement.js
@@ -7,7 +7,7 @@ import Grid from '@material-ui/core/Grid';
 import Link from '@material-ui/core/Link';
 import Button from '@material-ui/core/Button';
 
-import bg_image from './bg1.jpeg';
+
 
 const useStyles = makeStyles((theme) => ({
   mainFeaturedPost: {
