+++ src/components/Header.js
@@ -3,15 +3,13 @@ import PropTypes from 'prop-types';
 import { makeStyles } from '@material-ui/core/styles';
 import Toolbar from '@material-ui/core/Toolbar';
 import Button from '@material-ui/core/Button';
-import IconButton from '@material-ui/core/IconButton';
 import Typography from '@material-ui/core/Typography';
-import Link from '@material-ui/core/Link';
 import SupervisedUserCircleIcon from '@material-ui/icons/SupervisedUserCircle';
-
+import Link from '@material-ui/core/Link';
 
 const useStyles = makeStyles((theme) => ({
   toolbar: {
-    borderBottom: `1px solid ${theme.palette.divider}`,
+    borderBottom: `2px solid ${theme.palette.divider}`,
   },
   toolbarTitle: {
     flex: 1,
@@ -43,10 +41,11 @@ export default function Header(props) {
           noWrap
           className={classes.toolbarTitle}
         >
-          {props.page_title}
+          <Link href="/"  color="inherit" >{props.page_title}</Link>
+          
         </Typography>
 
-        <Button color="secondary"> 
+        <Button color="secondary" href="/signin"> 
           Sign in
         </Button>
