+++ src/components/Landing.js
@@ -1,26 +1,12 @@
 import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import CssBaseline from '@material-ui/core/CssBaseline';
 import Grid from '@material-ui/core/Grid';
-import Container from '@material-ui/core/Container';
-import GitHubIcon from '@material-ui/icons/GitHub';
-import FacebookIcon from '@material-ui/icons/Facebook';
-import TwitterIcon from '@material-ui/icons/Twitter';
-import Header from './Header';
-import MainLandingElement from './MainLandingElement'
-import Featured from './Featured';
-import Footer from './Footer';
-
+import MainLandingElement from '../MainLandingElement'
+import Featured from '../Featured';
 
 import bg_image from './bg1.jpeg';
 import ana from './ana.png';
 import jack from './jack.png';
 
-const useStyles = makeStyles((theme) => ({
-  mainGrid: {
-    marginTop: theme.spacing(3),
-  },
-}));
 
 const mainLandingElementContent = {
   title: 'Connecting people in difficult times',
@@ -50,23 +36,12 @@ const featured = [
   },
 ];
 
-const social_icons = [
-    { name: 'GitHub', icon: GitHubIcon },
-    { name: 'Twitter', icon: TwitterIcon },
-    { name: 'Facebook', icon: FacebookIcon },
-  ]
-
+// todo ES6 style
 
 export default function Landing() {
-  const classes = useStyles();
-
   return (
     <React.Fragment>
-      <CssBaseline />
-      <Container maxWidth="lg">
-        <Header page_title="Hand in Hand" />
-        <main>
-          <MainLandingElement post={mainLandingElementContent} />
+      <MainLandingElement post={mainLandingElementContent} />
           
           <Grid container spacing={4}>
 
@@ -75,9 +50,8 @@ export default function Landing() {
             ))}
 
           </Grid>
-        </main>
-      </Container>
-      <Footer title="HTG2020" description="Something here to give the footer a purpose!"  social={social_icons} />
+
+      
     </React.Fragment>
   );
 }
